requirements.txt
flask==3.0.0
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.6.0
marshmallow==3.20.1
python-dotenv==1.0.0
sqlalchemy
pytest==7.4.3
gunicorn==21.2.0
psycopg[binary]
supabase
